Update and rename group_func.py to word_cloud.py
